requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.4.0
exceptiongroup==1.1.3
Faker==23.2.1
flake8==7.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
setuptools==69.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.9.0
urllib3==2.2.1
webdriver-manager==4.0.1
wheel==0.43.0
wsproto==1.2.0

